v0.3.5, fixed 0.3.4's adjustment of microsecond timing

+++ src/ARGB.cpp
@@ -3,7 +3,7 @@
 #include "ARGB.h"
 
 
-ARGB::ARGB(int32_t ledCount) {
+ARGB::ARGB(uint8_t ledCount) {
 	// Save the given LED count
 	ARGB::ledCount = ledCount;
